Split compiler lineage helpers
